# Update xml char type