--- a/src/libraries/System.Private.Xml/src/System/Xml/BinaryXml/XmlBinaryReader.cs
+++ b/src/libraries/System.Private.Xml/src/System/Xml/BinaryXml/XmlBinaryReader.cs
@@ -333,7 +333,6 @@ namespace System.Xml
         private readonly bool _ignoreComments;
         private readonly DtdProcessing _dtdProcessing;
 
-        private XmlCharType _xmlCharType;
         private readonly Encoding _unicode;
 
         // current version of the protocol
@@ -342,7 +341,6 @@ namespace System.Xml
         public XmlSqlBinaryReader(Stream stream, byte[] data, int len, string baseUri, bool closeInput, XmlReaderSettings settings)
         {
             _unicode = System.Text.Encoding.Unicode;
-            _xmlCharType = XmlCharType.Instance;
 
             _xnt = settings.NameTable!;
             if (_xnt == null)
@@ -3531,7 +3529,6 @@ namespace System.Xml
             // assert that size is an even number
             Debug.Assert(0 == ((_pos - _tokDataPos) & 1), "Data size should not be odd");
             // grab local copy (perf)
-            XmlCharType xmlCharType = _xmlCharType;
 
             fixed (byte* pb = _data)
             {
@@ -3546,7 +3543,7 @@ namespace System.Xml
                         int posNext = pos + 2;
                         if (posNext > end)
                             return _xmlspacePreserve ? XmlNodeType.SignificantWhitespace : XmlNodeType.Whitespace;
-                        if (pb[pos + 1] != 0 || !xmlCharType.IsWhiteSpace((char)pb[pos]))
+                        if (pb[pos + 1] != 0 || !XmlCharType.IsWhiteSpace((char)pb[pos]))
                             break;
                         pos = posNext;
                     }
@@ -3561,7 +3558,7 @@ namespace System.Xml
                         if (posNext > end)
                             return XmlNodeType.Text;
                         ch = (char)(pb[pos] | ((int)(pb[pos + 1]) << 8));
-                        if (!_xmlCharType.IsCharData(ch))
+                        if (!XmlCharType.IsCharData(ch))
                             break;
                         pos = posNext;
                     }