--- a/src/libraries/System.Private.Xml/src/System/Xml/Core/XmlCharCheckingWriter.cs
+++ b/src/libraries/System.Private.Xml/src/System/Xml/Core/XmlCharCheckingWriter.cs
@@ -24,7 +24,6 @@ namespace System.Xml
         private readonly bool _replaceNewLines;
         private readonly string _newLineChars;
 
-        private XmlCharType _xmlCharType;
         //
         // Constructor
         //
@@ -37,10 +36,6 @@ namespace System.Xml
             _replaceNewLines = replaceNewLines;
             _newLineChars = newLineChars;
 
-            if (checkValues)
-            {
-                _xmlCharType = XmlCharType.Instance;
-            }
         }
 
         //
@@ -83,7 +78,7 @@ namespace System.Xml
                 if (pubid != null)
                 {
                     int i;
-                    if ((i = _xmlCharType.IsPublicId(pubid)) >= 0)
+                    if ((i = XmlCharType.IsPublicId(pubid)) >= 0)
                     {
                         throw XmlConvert.CreateInvalidCharException(pubid, i);
                     }
@@ -238,7 +233,7 @@ namespace System.Xml
             if (_checkNames)
             {
                 int i;
-                if ((i = _xmlCharType.IsOnlyWhitespaceWithPos(ws)) != -1)
+                if ((i = XmlCharType.IsOnlyWhitespaceWithPos(ws)) != -1)
                 {
                     throw new ArgumentException(SR.Format(SR.Xml_InvalidWhitespaceCharacter, XmlException.BuildCharExceptionArgs(ws, i)));
                 }