… src/libraries/System.Private.Xml/src/System/Xml/Core/XmlTextReaderImpl.cs
… src/libraries/System.Private.Xml/src/System/Xml/Core/XmlTextReaderImplAsync.cs
… src/libraries/System.Private.Xml/src/System/Xml/Core/XmlUtf8RawTextWriterAsync.cs
… src/libraries/System.Private.Xml/src/System/Xml/Core/XmlWellFormedWriter.cs
… src/libraries/System.Private.Xml/src/System/Xml/Core/XmlWellFormedWriterAsync.cs
… src/libraries/System.Private.Xml/src/System/Xml/Core/XmlWellFormedWriterHelpers.cs
… src/libraries/System.Private.Xml/src/System/Xml/Dom/XmlCharacterData.cs
… src/libraries/System.Private.Xml/src/System/Xml/Schema/DtdParser.cs
… src/libraries/System.Private.Xml/src/System/Xml/Schema/DtdParserAsync.cs
… src/libraries/System.Private.Xml/src/System/Xml/Schema/Parser.cs
… src/libraries/System.Private.Xml/src/System/Xml/Schema/XmlSchemaValidator.cs
… src/libraries/System.Private.Xml/src/System/Xml/ValidateNames.cs
… src/libraries/System.Private.Xml/src/System/Xml/XPath/Internal/StringFunctions.cs
… src/libraries/System.Private.Xml/src/System/Xml/XPath/Internal/XPathScanner.cs
… src/libraries/System.Private.Xml/src/System/Xml/XmlCharType.bin
… src/libraries/System.Private.Xml/src/System/Xml/XmlCharType.cs
… src/libraries/System.Private.Xml/src/System/Xml/XmlComplianceUtil.cs
… src/libraries/System.Private.Xml/src/System/Xml/XmlConvert.cs
… src/libraries/System.Private.Xml/src/System/Xml/Xsl/Runtime/WhitespaceRuleReader.cs
… src/libraries/System.Private.Xml/src/System/Xml/Xsl/Runtime/XsltFunctions.cs
… src/libraries/System.Private.Xml/src/System/Xml/Xsl/XPath/XPathScanner.cs
… src/libraries/System.Private.Xml/src/System/Xml/Xsl/XslException.cs
… src/libraries/System.Private.Xml/src/System/Xml/Xsl/Xslt/XsltInput.cs
… src/libraries/System.Private.Xml/src/System/Xml/Xsl/XsltOld/BuilderInfo.cs
… src/libraries/System.Private.Xml/src/System/Xml/Xsl/XsltOld/Compiler.cs
… src/libraries/System.Private.Xml/src/System/Xml/Xsl/XsltOld/ProcessingInstructionAction.cs
… src/libraries/System.Private.Xml/src/System/Xml/Xsl/XsltOld/ReaderOutput.cs
… src/libraries/System.Private.Xml/src/System/Xml/Xsl/XsltOld/SequentialOutput.cs