fixes https://github.com/dotnet/runtime/issues/41572

This removes an embedded resource binary file along with the support code needed to load it safely and atomically. It is replaced with an enormous readonly span which is inlined into the data segment by the compiler. Doing this allows the struct to be safe, not need a ctor and for the static Instance method to be default(T), the entire thing could probably be static

The changes are split into two commits. The first has the code changes, review this. The second only contains the span which was added to the bottom of the file so the contents can be seen casually without forcing a scroll through almost 4000 lines of code. The span code was generated from the binary in the project, review it only if you like hex.

/cc @RUSshy