@tarekgh 2020-09-02
>             /* 000 */ 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x11, 0x11, 0x00, 0x00, 0x11, 0x00, 0x00,

nit: could you please have this in hex instead as the table is going with 16 elements in row? and if you can add a column header like 

// 0  1  2 3  ... A B C D E F

just to make it easy to lookup any character

---
Refers to: src/libraries/System.Private.Xml/src/System/Xml/XmlCharType.cs:233 in bfd86e5. [](commit_id = bfd86e593f1fc2c26e37f630f92cd820455b7256, deletion_comment = False)