@tarekgh 2020-09-02
@Wraith2 thanks for making this change. LGTM. 

@krwq @buyaa-n maybe the next step here is to review https://www.w3.org/TR/xml/ and ensure this data we carry is conforming to the latest Xml specs.