@Wraith2 2020-09-02
> I assume you have converted the bin file directly to the span without changing any value. I looked at some random rows and I am not seeing any change

Yes. I wrote a small program to read the file 16 bytes at a time and output the bytes in code format. No space for human error.