@tarekgh 2020-09-02
> Trying to anything with the file that involves a format change will max the compiler for a while so be careful.

Maybe later we can try to convert this table into a double lookup table (if not affecting the perf much) which may make the tables much smaller. This will need some time to do perf measurement and generate the optimized tables. 
currently having a 64k table is really big and I admire the compiler it can handle it.