@danmosemsft 2020-09-03
> maybe the next step here is to review https://www.w3.org/TR/xml/ and ensure this data we carry is conforming to the latest Xml specs.

@Wraith2 is that something you're interested in doing? Or maybe @tarekgh did not mean that should hold up this PR.