@tarekgh 2020-09-03
> Or maybe @tarekgh did not mean that should hold up this PR.

This should not be a blocker this PR. @Wraith2 is not changing anything from what we already have. my comment was about to review the data we have and ensure it is conforming to the latest Xml spec.