@Wraith2 2020-09-03
> @Wraith2 is that something you're interested in doing? 

Not really not... Reviewing 64K of byte values individually while the compiler struggles to parse the data sounds like a punishment for both my computer and I. 😁