@tarekgh 2020-09-03
> Not really no... Reviewing 64K of byte values individually while the compiler struggles to parse the data sounds like a punishment for both my computer and I. 😁

@Wraith2 you can review everyday just 1k and we'll be done in couple of months :-) just kidding. thanks again for getting this change. 

CC @jkotas @davidwrighton if they have any concern having 64k readonly span data like that.