@jkotas 2020-09-03
This is modifying perf critical XML parsing code. Do we have any XML parsing microbenchmark to check the impact?