@tarekgh 2020-09-03
aside from the perf, I like the change because it allow checking any character properties by just looking at the array value. but in general, I agree, having perf numbers is better here.