@davidwrighton 2020-09-04
I looked at this, and I have no concerns about having a large `ReadOnlySpan<byte>` used here. However, I do think we should measure the performance of this, as this might be slower than the previous implementation. I doubt it, but I've been unpleasantly surprised before.

There are two possibilities for reduced performance that I can see here.
1. Accessing the span may be injecting a bounds check. In this case bounds check isn't actually needed, as the data is sufficiently large, but we might be missing that optimization in this case.
2. Accessing the span as embedded data requires embedding a pointer into the code stream of the jit. Reading that pointer from the code stream may be slower than reading a field from the struct. I don't expect that this is true, but on some platforms (notably arm64) it requires 2 instructions to embed a pointer where the field reading logic is a single instruction, and on intel architecture platforms, the field reading instruction is slightly bigger. OTOH, reading a field is an extra memory load, which is generally considered to be more expensive to run than a typical instruction.

Now, given our system, I don't expect this to be slower, I actually expect the lack of an extra memory read to make this all slightly faster than the previous behavior.