@jkotas 2020-09-06
Here is a small targeted test:
```
using System;
using System.IO;
using System.Xml;

class Program
{
    static void Main(string[] args)
    {
        var ms = new MemoryStream();

        using (var w = XmlWriter.Create(ms))
        {
            w.WriteStartDocument();
            w.WriteStartElement("OutterElement");
            for (int i = 0; i < 10000; i++)
            {
                w.WriteStartElement("InnerElement");
                w.WriteEndElement();
            }
            w.WriteEndElement();
            w.WriteEndDocument();
        }

        for (; ; )
        {
            int start = Environment.TickCount;
            for (int i = 0; i < 1000; i++)
            {
                ms.Position = 0;
                using (var r = XmlReader.Create(ms))
                {
                    while (r.Read())
                    {
                    }
                }
            }
            int end = Environment.TickCount;
            Console.WriteLine(end - start);
        }
    }
}
```

It runs about ~900ms per iteration before this change and ~1200ms per iteration after this change.