@Wraith2 2020-09-06
I was going to work on a test this morning but thanks for doing that. 

I don't know how to determine what the speed change is caused by to see if it can be mitigated. With those results I don't believe it would be an acceptable regression so either I close or look at flowing the instance/static changes mentioned above out to see if the ability to call and possible inline static methods recovers the perf. I'd like to be gathered by knowledge on that one though so if you can give me some info on what tools and methods you'd use to diagnose the speed difference that would be good.