@jkotas 2020-09-06
> I don't know how to determine what the speed change is caused by 

You can use profiler such as perfview to get the idea what methods to look at. This type of change will require looking at the disassembly to identify the bottlenecks. If it is not your expertise, there are number of people with this expertise hanging around and they may be able to help you if you ask.

I took a cursory look when I wrote the test and the first problem I have noticed was a lack of inlining for `XmlCharType.IsNCNameSingleChar` call from `XmlTextReaderImpl.ParseElement`, so part of the fix is likely going to be to add  AggressiveInlining on the XmlCharType accessors. It will fix some of the regression, but I suspect that it won't be enough.