@Wraith2 2020-09-06
I've got a local build of runtime that includes the change but I can't get BDN to use it at the moment, it seems to be ignoring the --coreRun argument entirely.

I used sharplab to have a look at the il and asm that is generated for the IsWhitespace call and it was surprising and given the length I'm not surprised it won't inline. If I can get to a position where I can replicate your bench I can try making it a ref struct and passing the span is as a ctor arg the way it was done previously which should cut down iteration cost.