@buyaa-n 2020-09-08
I have tried with the existing [benchmark ](https://github.com/dotnet/performance/blob/master/docs/benchmarking-workflow-dotnet-runtime.md) we have.  We don't have benchmarks exactly for reader/writer, `XmlDocument.LoadXml` could show the reader perf some level. But I don't understand why `XmlNode.GetAttributes` having so high regression. I might add some benchmarks for reader/writer and check the result.
| Slower                                                      | diff/base | Base Median (ns) | Diff Median (ns) | Modality|
| ----------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| XmlDocumentTests.XmlNodeTests.Perf_XmlNode.GetAttributes    |      1.66 |             0.64 |             1.07 |         |
| XmlDocumentTests.XmlDocumentTests.Perf_XmlDocument.LoadXml  |      1.07 |          2997.90 |          3203.79 |         |
| XmlDocumentTests.XmlNodeListTests.Perf_XmlNodeList.GetCount |      1.07 |            10.22 |            10.92 |         |